Refactor label printing functionality
- Move to helpers.dart
- Will be used for other label types also
