+++ lib/helpers.dart
@@ -9,12 +9,19 @@
 
 import "dart:io";
 import "package:currency_formatter/currency_formatter.dart";
+import "package:flutter/cupertino.dart";
+import "package:font_awesome_flutter/font_awesome_flutter.dart";
+import "package:inventree/api.dart";
+import "package:inventree/inventree/part.dart";
+import "package:inventree/widget/progress.dart";
 import "package:one_context/one_context.dart";
 import "package:url_launcher/url_launcher.dart";
 import "package:audioplayers/audioplayers.dart";
 
 import "package:inventree/l10.dart";
 import "package:inventree/widget/snacks.dart";
+
+import "api_form.dart";
 
 
 
@@ -132,4 +139,114 @@ String renderCurrency(double? amount, String currency, {int decimals = 2}) {
   }
 
   return value;
+}
+
+/*
+ * Select a particular label, from a provided list of options,
+ * and print against the selected instances.
+ */
+Future<void> selectAndPrintLabel(
+    BuildContext context,
+    List<Map<String, dynamic>> labels,
+    String labelType,
+    String labelQuery,
+) async {
+
+  // Find a list of available plugins which support label printing
+  var plugins = InvenTreeAPI().getPlugins(mixin: "labels");
+
+  dynamic initial_label;
+  dynamic initial_plugin;
+
+  List<Map<String, dynamic>> label_options = [];
+  List<Map<String, dynamic>> plugin_options = [];
+
+  // Construct list of available label templates
+  for (var label in labels) {
+    String display_name = (label["description"] ?? "").toString();
+    int pk = (label["pk"] ?? -1) as int;
+
+    if (display_name.isNotEmpty && pk > 0) {
+      label_options.add({
+        "display_name": display_name,
+        "value": pk,
+      });
+    }
+  }
+
+  if (label_options.length == 1) {
+    initial_label = label_options.first["value"];
+  }
+
+  // Construct list of available plugins
+  for (var plugin in plugins) {
+    plugin_options.add({
+      "display_name": plugin.humanName,
+      "value": plugin.key
+    });
+  }
+
+  if (plugin_options.length == 1) {
+    initial_plugin = plugin_options.first["value"];
+  }
+
+  Map<String, dynamic> fields = {
+    "label": {
+      "label": L10().labelTemplate,
+      "type": "choice",
+      "value": initial_label,
+      "choices": label_options,
+      "required": true,
+    },
+    "plugin": {
+      "label": L10().pluginPrinter,
+      "type": "choice",
+      "value": initial_plugin,
+      "choices": plugin_options,
+      "required": true,
+    }
+  };
+
+  launchApiForm(
+    context,
+    L10().printLabel,
+    "",
+    fields,
+    icon: FontAwesomeIcons.print,
+    onSuccess: (Map<String, dynamic> data) async {
+      int labelId = (data["label"] ?? -1) as int;
+      String pluginKey = (data["plugin"] ?? "") as String;
+
+      if (labelId != -1 && pluginKey.isNotEmpty) {
+        String url = "/label/${labelType}/${labelId}/print/?${labelQuery}&plugin=${pluginKey}";
+
+        showLoadingOverlay(context);
+
+        InvenTreeAPI().get(url).then((APIResponse response) {
+          hideLoadingOverlay();
+          if (response.isValid() && response.statusCode == 200) {
+
+            var data = response.asMap();
+
+            if (data.containsKey("file")) {
+              var label_file = (data["file"] ?? "") as String;
+
+              // Attempt to open remote file
+              InvenTreeAPI().downloadFile(label_file);
+            } else {
+              showSnackIcon(
+                  L10().printLabelSuccess,
+                  success: true
+              );
+            }
+          } else {
+            showSnackIcon(
+              L10().printLabelFailure,
+              success: false,
+            );
+          }
+        });
+      }
+    },
+  );
 }
